Round and align allocations allowing precise bounds.
The core change is to pad the front of the allocation so the allocation
is shifted up as required.  To get back to the beginning I just added
a pointer to the allocation (a pad sized could also be used, but I'm not
particularly concerned with efficency here.)  I've also removed the
end guard word.

(cherry picked from commit db0b6b9bfc684dac2b5f1e9208e8c777b5979814)