[cheriabi] Align portion of Parse struct
The function `sqlite3NestedParse` saves and restores a portion of the
`Parse` struct between the field `sLastToken` and the end of the struct
(using a `memcpy` to a temporary buffer). As this portion of the Parse
struct contains capability values the `memcpy`s that saved and restore
the struct must occur on memory aligned to `alignof(max_align_t)` (which
is equivalent to aligning to `SQLITE_DEFAULT_ALIGNMENT` for the
CheriABI).